// TM8/tm8-game-be-main/src/common/constants/game-preference-value.enum.ts
export enum GamePreferenceValue {
  Arena = 'arena',
  Duo = 'duo',
  Intermediate = 'intermediate',
  Aggressive = 'aggressive',
  TeamPlayer = 'team-player',
  FindPeople = 'find-people',
  CenterCircle = 'center-circle',
  Casual = 'casual',
  Hoops = 'hoops',
  TwoVTwo = '2v2',
  Beginner = 'beginner',
  PlayingLevel = 'playing-level',
  HowYouPlay = 'how-you-play',
  Mixtape = 'mixtape',
  Control = 'control',
  BattleRoyale = 'battle-royale',
  BattleRoyaleBuild = 'battle-royale-build',
  BattleRoyaleNoBuild = 'battle-royale-no-build',
  Advanced = 'advanced',
  CreativeMode = 'creative-mode',
  Trio = 'trio',
  Squad = 'squad',
  FirstTime = 'first-time',
  Competitive = 'competitive',
  Pro = 'pro',
  Legend = 'legend',
  ChaseKill = 'chase-kill',
  ZoneWalk = 'zone-walk',
  Resurgence = 'resurgence',
  Plunder = 'plunder',
  GunGame = 'gun-game',
  Ranked = 'ranked',
  Tournaments = 'tournaments',
  SnowDay = 'snow-day',
  Dropshot = 'dropshot',
  Rumble = 'rumble',
  ThreeVThree = '3v3',
  FourVFour = '4v4',
  Training = 'training',
  FiringRange = 'firing-range',
  RankedLeague = 'ranked-league',
  Assault = 'assault',
  Skirmisher = 'skirmisher',
  Recon = 'recon',
  Controller = 'controller',
  Support = 'support',
  TeamDeathmatch = 'team-deathmatch',
  RunGun = 'run-gun',
  Scorer = 'scorer',
  Offensive = 'offensive',
  Extras = 'extras',
  FiringSquad = 'firing-squad',
  DemoHeavy = 'demo-heavy',
  Passer = 'passer',
  Anchor = 'anchor',
  OneVOne = '1v1',
  OffensiveScorer = 'offensive-scorer',
  OffensivePasser = 'offensive-passer',
  MidScorer = 'mid-scorer',
  MidDefender = 'mid-defender',
  DefensiveScorer = 'defensive-scorer',
  DefensivePasser = 'defensive-passer',
  Warzone = 'warzone',
  QuickGames = 'quick-games',
  StartingTimestamp = 'starting-timestamp',
  EndingTimestamp = 'ending-timestamp',
}